SpanAnalyzerDoc changed container type to lists for storing weathercases and spans. This makes it easier to move/sort without destroying/creating objects. Since weathercases are referenced by pointer, this is very important. To replace the index parameter for accessing/modifying weathercases and spans, const_iterators are now used.
The WeathercaseTreeCtrl and SpanTreeCtrl have been updated to interact with
the document by using iterators instead of indices. To help with this, the
data associated with each treectrl item has been modified to store a
const_iterator to the object that it represents. These treectrls have also
been updated with drag and drop sorting.

Remaining XmlHandlers, and various other panels have been adapted to use the
list container, and iterators instead of indices as well.