Reorganized and added more functionality to the ResultsPane. Turned the single panel into a notebook to hold 2 addition panels. The SpanPanel contains the original panel, and the SagTensionTablePanel and the CatenaryTablePanel were added.
The results pane now calculates and caches sag-tension values for all
weathercases of the activated span. Each individual pane then interprets
the sag-tension data and displays in its own unique way.

The SpanPanel was updated to interface with the calcuated values, instead
of calculating them with it's own LineCableReloader. Although unique catenary
spacing cannot be specified now, it will be a future feature.

The SagTensionTablePanel shows the sag-tension results in table form. Only
minimal catenary data is provided in this view, but instead focuses on tension
of the cable and components.

The CatenaryTable shows the catenary results in table form. This table expands
and complements the SagTensionTablePanel. Although unique catenary spacing
cannot be specified now, it will be a future feature.