Updated application config and application-wide data handling. The SpanAnalyzerData now has it's own handler which loads data from its own file. The SpanAnalyzerData now contains weathercase sets, which are groups of weathercases specific to the analysis being performed.
Renamed the application config file, and moved some of the functionality to
the SpanAnalyzerData handler.

Renamed and updated the PreferencesDialog to handle an application data file
path.

Updated the ResultsPane to include a wxChoice control to determine which
weathercase set is to be analyzed. The dependent panels now grab the
weathercases selected by the ResultsPane choice.