Updated FileHandler to load/save cable files. FileHandler was also updated to convert unit system and style to match the application preferences.
Removed load cable files from SpanAnalyzerDataXmlHandler, as this
was added to the FileHandler instead.

Added a CableDirectoryEditorDialog to allow user to specify cable
directory. This seemed like the easiest approach, rather than a
manager dialog. A manager dialog would have been complicated because
each cable is stored in a separate file.

Updated SpanAnalyzerFrame menu to handle cable editing. The cable
directory can be edited, which affects which cables are loaded into
application memory, and standalone cable files can be edited as well.
Removed the debug menu event because the program is mature enough
that it isn't used very often. The frame interacts with the
FileHandler, which simplifies loading/saving.