Fixed the app config file saving on close.
Instead of saving the file when the frame closes, the frame
saves the size in the config settings and the file is saved
in the app exit function.

The config file doesn't say "Manual" or "Automatic" for the frame
size xml node content anymore, itjust stores the size as
attributes.

When the app starts, if the frame height or width dips below 100,
it'll trigger it to be shown maximized.