Added a 'Default' weathercase group to the data file.
The document weathercases are going to be replaced by the default
weathercase group. This is being done to simplify the document,
and so the user doesn't have to input weathercases on every new
file. All of the data referenced by the document spans will
eventually reference the application data, not withiin the
document itself. This may make the file more prone to 'breaking'
if reference data is missing, but this is seen as a reasonable
compromise for usability.