Changed how weathercases are stored in the application.
It all started with a bug. When the weathercases are sorted or
rearranged, the memory address does not stay fixed for a given
weathercase. This causes the weathercases referenced by the
spans to change to a different value. To fix this the
weathercases need to be manually allocated, and the list of
weathercases need to be pointers to the static memory address.

The application is designed to store several lists of
weathercases. To avoid manually managing the memory of several
lists, the weathercase list is condensed to a single one. This
causes significant application modifications.

I'm also going to cheat and add another significant change into
this commit. AnalysisFilters were created to display combinations
of weathercases and cable conditions. Now instead of creating
weathercase groups, analysis filter groups are chosen instead.
This allows the user to view results that have mixed conditions,
which required toggling a control before. This is helpful for
looking at max sag, where the initial/final condition can both be
viewed at the same time.

- Changed SpanAnalyzerData to store a list of WeatherLoadCase
  pointers.
- Created an AnalysisFilter class to pair WeatherLoadCase and
  CableConditionType.
- Updated XmlHandlers to accept list of WeatherLoadCase pointers.
- Updated EditorDialogs to accept list of WeatherLoadCase pointers.
- Updated SagTensionAnalysisResult struct to have references to
  the weathercase and condition.
- Updated AnalysisController to calculate results for the single
  weathercase list using all cable conditions. Changed results
  so they are stored as vectors. This shouldn't matter because
  the data isn't ever rearranged without deleting everything
  first. Modified the accessors to return pointers based on
  weathercase and condition combination.
- Updated SpanAnalyzerDoc. Changed hint update enumerations.
  Changed accessors to match AnalysisController.
- Updated SpanAnalyzerView to track the active analysis filter
  group and index.
- Updated ResultsPane and how it converts results from the
  document to the report table. Replaced the weathercase group
  with a filter group choice control, and removed the condition
  choice control.
- Updated PlotPane and how it converts results from the document
  to the plot.
- Added an AnalysisFilterManagerDialog class to allow user to edit
  AnalysisFilterGroups.
- Added an AnalysisFilterEditorDialog class to allow user to edit
  AnalysisFilters.
- Modified WeatherLoadCaseManagerDialog to edit a single list of
  WeatherLoadCases. This dialog will manually manage/allocate
  memory as needed when creating new weathercases.
- Updated SpanAnalyzerFrame to have a menu item for launching the
  AnalysisFilterManagerDialog.
- Updated test files and appdata.xml.